Update README.md
SARathi is a Python-based, AI-driven pipeline for intelligent satellite image analysis. It leverages Synthetic Aperture Radar (SAR) imagery and Generative AI to provide actionable insights for critical applications like agriculture, disaster management, and environmental monitoring. The pipeline is designed for efficiency and is built with a modular architecture, making it highly scalable for future applications.